requirements.txt
# oslo==1.0.1 https://github.com/cuteboy9201/oslo.git 
tornado==4.3
sqlalchemy==1.3.10
pymysql==0.9.3
pyjwt==1.7.1
pillow==6.2.1